one split-first-rows step, from streaming and parquet
