+++ libs/libcommon/tests/test_processing_graph.py
@@ -79,7 +79,7 @@ def test_graph() -> None:
             "config-split-names",
             [
                 "config-opt-in-out-urls-count",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "dataset-split-names",
                 "split-duckdb-index",
                 "split-duckdb-index-010",
@@ -104,21 +104,15 @@ def test_graph() -> None:
             ],
         ),
         (
-            "split-first-rows-from-parquet",
+            "split-first-rows",
             ["split-is-valid", "split-image-url-columns"],
-            ["config-parquet-metadata"],
-            ["config-parquet", "dataset-config-names", "config-parquet-and-info", "config-parquet-metadata"],
-        ),
-        (
-            "split-first-rows-from-streaming",
-            ["split-is-valid", "split-image-url-columns"],
+            ["config-parquet-metadata", "config-split-names"],
             [
-                "config-split-names",
-            ],
-            [
+                "config-parquet",
                 "dataset-config-names",
                 "config-split-names",
                 "config-parquet-and-info",
+                "config-parquet-metadata",
                 "config-info",
             ],
         ),
@@ -130,11 +124,7 @@ def test_graph() -> None:
         ),
         (
             "config-parquet-metadata",
-            [
-                "split-first-rows-from-parquet",
-                "split-duckdb-index",
-                "split-duckdb-index-010",
-            ],
+            ["split-first-rows", "split-duckdb-index", "split-duckdb-index-010"],
             ["config-parquet"],
             ["dataset-config-names", "config-parquet-and-info", "config-parquet"],
         ),
@@ -189,8 +179,7 @@ def test_graph() -> None:
                 "config-size",
                 "config-split-names",
                 "config-parquet-metadata",
-                "split-first-rows-from-parquet",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "config-is-valid",
                 "split-is-valid",
                 "split-duckdb-index",
@@ -199,16 +188,15 @@ def test_graph() -> None:
         (
             "split-image-url-columns",
             ["split-opt-in-out-urls-scan"],
-            ["split-first-rows-from-streaming", "split-first-rows-from-parquet"],
+            ["split-first-rows"],
             [
                 "dataset-config-names",
                 "config-split-names",
                 "config-info",
                 "config-parquet-and-info",
                 "config-parquet-metadata",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "config-parquet",
-                "split-first-rows-from-parquet",
             ],
         ),
         (
@@ -221,9 +209,8 @@ def test_graph() -> None:
                 "config-info",
                 "config-parquet-and-info",
                 "config-parquet-metadata",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "config-parquet",
-                "split-first-rows-from-parquet",
                 "split-image-url-columns",
             ],
         ),
@@ -234,13 +221,12 @@ def test_graph() -> None:
             [
                 "dataset-config-names",
                 "config-split-names",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "config-info",
                 "config-parquet-and-info",
                 "config-parquet-metadata",
                 "split-opt-in-out-urls-scan",
                 "config-parquet",
-                "split-first-rows-from-parquet",
                 "split-image-url-columns",
             ],
         ),
@@ -251,14 +237,13 @@ def test_graph() -> None:
             [
                 "dataset-config-names",
                 "config-split-names",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "config-info",
                 "config-parquet-and-info",
                 "config-parquet-metadata",
                 "split-opt-in-out-urls-count",
                 "split-opt-in-out-urls-scan",
                 "config-parquet",
-                "split-first-rows-from-parquet",
                 "split-image-url-columns",
             ],
         ),
@@ -269,7 +254,7 @@ def test_graph() -> None:
             [
                 "dataset-config-names",
                 "config-split-names",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "config-info",
                 "config-parquet-and-info",
                 "config-parquet-metadata",
@@ -277,7 +262,6 @@ def test_graph() -> None:
                 "split-opt-in-out-urls-count",
                 "split-opt-in-out-urls-scan",
                 "config-parquet",
-                "split-first-rows-from-parquet",
                 "split-image-url-columns",
             ],
         ),
@@ -352,8 +336,7 @@ def test_graph() -> None:
                 "dataset-loading-tags",
                 "dataset-size",
                 "split-duckdb-index",
-                "split-first-rows-from-parquet",
-                "split-first-rows-from-streaming",
+                "split-first-rows",
                 "split-is-valid",
             ],
         ),
