+++ services/worker/tests/job_runners/split/test_image_url_columns.py
@@ -194,7 +194,7 @@ def test_compute(
         app_config,
     )
     upsert_response(
-        kind="split-first-rows-from-streaming",
+        kind="split-first-rows",
         dataset=dataset,
         config=config,
         split=split,
@@ -239,7 +239,7 @@ def test_compute_failed(
     )
     if dataset != "doesnotexist":
         upsert_response(
-            kind="split-first-rows-from-streaming",
+            kind="split-first-rows",
             dataset=dataset,
             config=config,
             split=split,
